Async profiler: CPU vs IO example.
By default, Async profiler does (on-)CPU profiling
so it cannot construct a proper profile for threads blocked on IO.

We demonstrate this by trying a bunch of network calls that take about 10 seconds
and also doing some CPU intensive operations for 2 seconds.
The resulting profile is misleading because it shows that the major part
is actually `burn-cpu` while it's less than 20% of the total running time.